Bitstream fix for reading arrays; Megalo value-type constexpr fixes
Fixed cobb::bitstream::read so that it now works with bounded arrays e.g. uint32_t[5]. Added a template-helper macro (cobb_enable_case) to effect this change. Altered BlamHeader to make use of this (TODO: other stuff, maybe the SHA1 hash in MPVR).

Renamed cobb::enum_or_int_to_int to cobb::strip_enum.

Added documentation comments to cobb::bitnumber.

Basically all of the Megalo opcode value stuff is constexpr now, and the project can build again.

+++ native/src/ReachVariantTool/helpers/files.h
@@ -43,7 +43,7 @@ namespace cobb {
          }
          //
          inline bool is_in_bounds(uint32_t offset, uint32_t bytes) const noexcept {
-            return (offset + bytes) < this->size;
+            return ((uint64_t)offset + bytes) < this->size;
          }
    };
    class file_guard {
